% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darflex.r
\docType{package}
\name{darflex-package}
\alias{darflex}
\alias{darflex-package}
\title{Tools to process DARDAR and FLEXPART}
\description{
Here we just load dplyr so that the \%>\% operator is recognized throughout
}
\author{
\strong{Maintainer}: Odran Sourdeval \email{odran.sourdeval@univ-lille.fr}

}
