% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_origin.r
\name{find_origin_overpass}
\alias{find_origin_overpass}
\title{Determine Final Cloud Origin and Quality for FLEXPART Trajectories}
\usage{
find_origin_overpass(df_in)
}
\arguments{
\item{df_in}{A data frame containing the input trajectory data. This should include variables such as \code{dt_ice},
\code{always_cold}, \code{always_warm}, and others required for origin determination.}
}
\value{
A data frame with particle trajectories and their final cloud origin (\code{origin}), along with associated cloud parameters
(latitude, longitude, temperature) and a quality flag (\code{origin_quality}).
}
\description{
This function refines the origin of cloud formation in FLEXPART trajectories using a series of case conditions.
It assigns a final origin (\code{origin_final}) and a quality score based on temperature and trajectory data.
The output includes the final origin, associated cloud data (latitude, longitude, temperature), and quality flags.
}
\details{
The function applies several case conditions to assess the origin of clouds and the reliability of the origin identification.
It uses pre-determined conditions such as \code{always_cold}, \code{always_warm}, and \code{dt_ice} to adjust the final origin and quality flag.
}
