% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_origin.r
\name{find_origin_wernli}
\alias{find_origin_wernli}
\title{Identify Ice Cloud Origin and Characteristics in FLEXPART Trajectories}
\usage{
find_origin_wernli(df_in)
}
\arguments{
\item{df_in}{A data frame containing FLEXPART trajectory data, including cloud and temperature variables such as \code{flag_ice},
\code{origin_tmp}, \code{time_traj}, \code{lat}, \code{lon}, and \code{ta}.}
}
\value{
A data frame summarizing the origin of the ice cloud for each particle,
including the time difference between ice formation and overpass (\code{dt_cloud}),
the time difference between the last ice cloud and overpass (\code{dt_ice}), and information on whether the trajectory
remained in a cold or warm region. It also provides the latitude, longitude, and temperature at the point of ice formation.
}
\description{
This function identifies the origin of ice formation in FLEXPART trajectories by analyzing temperature and cloud flags.
It determines the time, location, and temperature at the point of ice formation and calculates the time difference between
the formation and the overpass.
}
\details{
The function uses temperature thresholds (e.g., below -40°C) to flag regions as cold and identifies when
ice formation occurs in the trajectory. The function summarizes the trajectory data to determine the origin of the ice
cloud and other related metrics.
}
