% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dardar_origin.r
\name{load_dardar_origin_2D}
\alias{load_dardar_origin_2D}
\title{Load and Process 2D DARDAR Origin Data}
\usage{
load_dardar_origin_2D(fn_dardar_origin, lf_dardar, lf_dardar_flux)
}
\arguments{
\item{fn_dardar_origin}{A character string specifying the file path to the DARDAR origin NetCDF file.}

\item{lf_dardar}{A character vector containing the list of available DARDAR NetCDF file paths.}

\item{lf_dardar_flux}{A character vector containing the list of available DARDAR CRE NetCDF file paths.}
}
\value{
A data frame containing combined cloud properties from the DARDAR and DARDAR origin files, including latitude, longitude, ice water path (IWP), ice concentration (ICNC), cloud top temperature (CTT), and cloud radiative effects (CRE).

A processed data frame with key cloud properties and radiative effects.
}
\description{
This function loads and processes 2D DARDAR origin and related DARDAR data, including DARDAR cloud radiative effect (CRE) data. It combines cloud properties such as ice water content (IWC), ice concentration, and temperature at cloud top, and merges it with DARDAR flux data (CRE).
}
\details{
The function attempts to match the DARDAR origin file with the corresponding DARDAR and DARDAR flux files based on the file name. It verifies that the dimensions of these files are consistent, then extracts and combines cloud-related data such as origin, IWC, ice concentration, temperature, and cloud radiative effects. Data is filtered based on cloud mask, mixed-phase flags, and iteration flags.
}
