% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.r
\name{load_data}
\alias{load_data}
\title{Load and Process FLEXPART and DARDAR Data}
\usage{
load_data(fn_flex, lf_dardar, dir_out, overwrite = FALSE, suffix = "")
}
\arguments{
\item{fn_flex}{A character string specifying the file path to the FLEXPART data.}

\item{lf_dardar}{A character vector of available DARDAR file paths.}

\item{dir_out}{A character string specifying the output directory where the processed file will be saved.}

\item{overwrite}{Logical, if \code{TRUE}, any existing output file with the same name will be overwritten. Default is \code{FALSE}.}

\item{suffix}{A character string appended to the output file name for distinguishing different runs. Default is an empty string.}
}
\value{
Returns \code{NULL}. The processed data is saved as a NetCDF file in the specified output directory.
}
\description{
This function loads FLEXPART and DARDAR data, computes the origin of particles from FLEXPART, and merges it with the DARDAR data. The combined data is then written to a new NetCDF file. Optionally, an existing output file can be overwritten.
}
\details{
The function identifies the correct DARDAR file corresponding to the FLEXPART file, computes particle origins using \code{find_origin_overpass}, and merges FLEXPART and DARDAR data at the time of overpass. The combined data is written to a new NetCDF file with the option to overwrite.
}
