% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_flex_dardar.r
\name{merge_dardar_flex}
\alias{merge_dardar_flex}
\title{Merge FLEXPART Overpass Data with DARDAR Data}
\usage{
merge_dardar_flex(df_flex_overpass, df_dardar)
}
\arguments{
\item{df_flex_overpass}{A data frame containing FLEXPART overpass data, including columns for height, origin, time, and other related variables.}

\item{df_dardar}{A data frame containing DARDAR data with columns for time, height, latitude, longitude, and other related variables.}
}
\value{
A merged data frame where DARDAR data is augmented with FLEXPART overpass data, binned by time and height. The merged data includes summary statistics such as origin standard deviation, mean origin, and cloud-related information for each bin.
}
\description{
This function merges FLEXPART overpass data with DARDAR data by binning the time and height variables into matching grids. It computes summary statistics for the FLEXPART data in each bin and joins it with the corresponding DARDAR data.
}
\details{
The function bins both FLEXPART and DARDAR data into specified intervals for time, height, latitude, and longitude. It then computes summary statistics (mean, standard deviation, etc.) for the FLEXPART data within each bin and merges it with the DARDAR data at the corresponding binned time and height levels.
}
