% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dardar.r
\name{nc_read}
\alias{nc_read}
\title{Read Variable from a NetCDF File}
\usage{
nc_read(nc, varname, idx = NULL)
}
\arguments{
\item{nc}{A NetCDF file object, typically opened with \code{ncdf4::nc_open}.}

\item{varname}{A character string specifying the name of the variable to be read from the NetCDF file.}

\item{idx}{An optional vector of indices to subset the variable. If \code{NULL} (default), the entire variable is read.}
}
\value{
A numeric vector containing the values of the specified variable from the NetCDF file, either entirely or by the specified indices.
}
\description{
This function reads a variable from a NetCDF file, either entirely or by specific indices, and returns it as a numeric vector.
}
\details{
The function uses \code{ncdf4::ncvar_get} to extract the variable's data from the NetCDF file. If no indices are provided, the entire variable is read. If indices are provided, only the specified subset is returned.
}
