% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_flexpart.r
\name{read_flexpart}
\alias{read_flexpart}
\title{Read FLEXPART Trajectory Data and Process Cloud and Temperature Variables}
\usage{
read_flexpart(fn)
}
\arguments{
\item{fn}{A character string specifying the path to the NetCDF file containing the FLEXPART trajectory data.}
}
\value{
A data frame with processed trajectory data, including time, particle, temperature, latitude, longitude, height, cloud flags (ice and liquid water), and a temporary origin identifier.
}
\description{
This function reads and processes FLEXPART trajectory data from a NetCDF file. It extracts key variables such as longitude, latitude, height, temperature, ice water content (IWC), and liquid water content (LWC). The function applies thresholds to identify clouds, converts temperatures from Kelvin to Celsius, and handles potential edge cases with the data.
}
\details{
\itemize{
\item The temperature is converted from Kelvin to Celsius.
\item Thresholds are applied to the IWC and LWC variables to identify clouds.
\item The function also accounts for a known issue with the first and last time steps in the data.
}
}
