% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_flexpart.r
\name{safe_open_ncdf}
\alias{safe_open_ncdf}
\title{Safely Open a NetCDF File}
\usage{
safe_open_ncdf(file_path)
}
\arguments{
\item{file_path}{A character string specifying the path to the NetCDF file to be opened.}
}
\value{
An object of class \code{ncdf4} if the file is successfully opened, or \code{NULL} if an error occurs.
}
\description{
This function attempts to open a NetCDF file using \code{ncdf4::nc_open}. If an error occurs (e.g., the file cannot be opened), it catches the error and returns \code{NULL} instead of stopping execution.
}
\details{
This function uses \code{tryCatch} to handle errors that may occur when opening a NetCDF file. If the file cannot be opened, an error message is printed, and the function returns \code{NULL}.
}
