% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dardar_flex.r
\name{write_dardar_flex}
\alias{write_dardar_flex}
\title{Write Combined DARDAR and FLEXPART Data to a NetCDF File}
\usage{
write_dardar_flex(df_dardar_flex, fn_dardar, fn_out)
}
\arguments{
\item{df_dardar_flex}{A data frame containing the combined DARDAR and FLEXPART data, with variables such as \code{n_part}, \code{origin}, \code{origin_quality}, \code{dt_cloud}, \code{ta_origin}, \code{lat_origin}, and \code{lon_origin}.}

\item{fn_dardar}{A character string specifying the file path to the original DARDAR NetCDF file, used to extract base variables such as latitude, longitude, and time.}

\item{fn_out}{A character string specifying the file path for the output NetCDF file. If the file exists, it will be overwritten.}
}
\value{
Returns \code{NULL}. The function creates a new NetCDF file at the specified location with the processed and combined data.
}
\description{
This function writes combined DARDAR and FLEXPART data to a new NetCDF file. It creates the necessary variables and dimensions based on the input data and writes the values for particle information, origin, cloud properties, temperature, and geographic information.
}
\details{
The function first reads the original DARDAR NetCDF file to extract dimensions and base variables. It then merges the data from \code{df_dardar_flex} into a full grid, fills missing values with a specified fill value, and creates the necessary variables in the NetCDF file. The new file contains variables for the number of particles, origin, origin quality, temperature at ice formation, latitude, and longitude, among others.
}
